complete coroutine system

+++ Assets/HQFramework/BaseFramework/Base/TaskSystem/ITaskDispatcher.cs
@@ -2,7 +2,7 @@ namespace HQFramework
 {
     public interface ITaskDispatcher<T> where T : TaskBase
     {
-        int MaxConcurrentCount
+        ushort MaxConcurrentCount
         {
             get;
         }
